Implemented emoji censor
To stop pointless emoji spam, the bot can now automatically delete any 
messages that qualify as 'emoji' spam. If there are only two characters 
of actual text, or the emojis equal or outnumber the characters of 
actual text, the bot will instantly delete the message without warning 
or prejudice. Unfortunately, this so far only works on Discord's custom 
emojis. the official Unicode emojis are currently unaffected.