Added YAG snipe
Built a new feature I call YAG Snipe. I wrote it because on Ponyworld, 
every time the YAGPDB.xyz bot is installed, it's always a recipe for 
disaster. This feature will immediately ban or kick the bot the moment 
it's added, and if it fails due to lack of permissions, it'll try again 
after any change that could possibly allow it to do so. It'll also try 
to kick the bot immediately after joining a new server. This feature can 
be repurposed merely by changing YAG_ID, to whatever bot or user one 
wishes to snipe on-sight.