Edited backup code alert message.
Noticed I was getting a flood of messages in the terminal from the bot, 
telling me it was "Running backup code in purgeActive". But it told be 
so little that the information was effectively useless. So I added some 
extra information to the message. The most useful thing being the time 
stamp.