Implemented payday command
The new payday command is the first step in a currency system for users
to take advantage of. What they will buy with this, I'm not sure yet.
Right now, all they can do is run the payday command, which will give
them 250 'bits'. It has a cooldown, which means they can only run it
once every 30 minutes. These figures can be adjusted. I made sure, as
well, to give detailed response and error messages, so the users know
what's going on.