Moved SQL file
There really is no need for an SQL folder with one file, and I highly 
doubt the database is going to be complex enough to warrant multiple 
files. So I moved the one SQL file out of the folder, renamed it to 
something more universal, and adjusted the readme to reflect the change.