Started work on rule management feature
Created a new rule management feature. This allows people to instantly 
store and lookup server rules on the fly. Without having to go to the 
'rules' channel and look them up. This is especially handy for mods 
trying to enforce the rules. They can quickly cite any rules that are 
being violated, with a single command.