Added two new rule commands
Two new commands for the server rule management function. All they do is 
output complete copies of the server rules to one's DMs, either in a 
standard human-readable form; or as a script one can use as an automated 
backup, so the rules can be recovered in the event of a database failure 
or in case the server needs to be rebuilt.