Dealt with commands in DMs
Bot was throwing exceptions whenever someone tried to send a command 
through DMs. So I fixed this by having the bot check if it was in a 
server before running most commands. It will still accept the help 
command in DMs, but I tweaked it so only the one command that works in 
DMs will appear in the command list.