Stop checkActive from running all the time
The checkActive code was running in every server, not just Ponyworld. 
And that's a problem because the hard-coded role id for the Active Role 
is only in Ponyworld. So running it anywhere else is pointless. Not 
really a major problem, in terms of server load, but I still think it 
was an oversight that needed to be corrected.