Added length control to rolecall
Had a bit of an issue with the rolecall function. In some cases, it was 
outputting blocks of text greater than 2000 characters. This is a 
problem because Discord's limit is 2000 characters. The only way I found 
to fix it was an if-else chain that tries to determine the best solution 
to the problem and fixes it by sending multiple messages. If none of the 
solutions fit (which would happen if there were more than 6000 
characters in a list of users with a particular top role) it outputs a 
message explaining this. I can't test all the parts of the chain, 
because I only have one server that triggers this problem, and I can't 
start shifting around members just to test a bot. It also doesn't fix 
the problem completely, because in some cases, it'll actually break the 
list in the middle of someone's name. But it's the best solution I have 
right now. I'll come up with a better one later. I also added TODO notes 
to remind me that I still need to write similar code for certain parts 
of the rule management feature. But it's a low-priority for now.