Altered on_error code
Was having a problem with the on_error code, designed to log exceptions, 
actually throwing exceptions. So I decided to try fixing this by having 
it catch and handle exceptions in a simple and logical way. Mainly by 
reporting to me, through DMs that it encountered a problem with the 
logging, while still reporting the exception to the console the 
old-fashioned way.