Forced UTF-8 encoding on error log
The error log wasn't always encoding in UTF-8 format, which was causing 
exceptions. This fixes that problem. It was why I wrote the code 
exception handling in the previous commit 
(f5c82317f851d6ab008750afd9fdbff5c9ce011f), which is a broader fix to 
deal with similar problems in the future. But this handles the immediate 
problem more reliably.