Rewrote length control in rolecall
The length control code from last night 
(6568608645215668da2096c408600cb8be9c9a04) was flawed in so many ways. 
It was breaking in the middle of names, wouldn't work if a particular 
role listed more than 6000 characters worth of names, and was just bulky 
and ridiculous. So I rewrote it to something simpler and more effective. 
It resolves to split the message on line breaks for a start and uses a 
much simpler design. It also scales much better. It's really the code I 
should've designed from the start.