Started moving active role settings to database
The settings for the active role feature were hard-coded in the program. 
This meant that it could really only work on one server, and the server 
administrator couldn't make any changes. Setting it up in the database 
meant that it could work on multiple servers, server admins could make 
little tweaks, and those tweaks could be different on different servers. 
So that's what I did. I created a new table to hold the settings in the 
database, and rewrote the checkActive function to pull the settings from 
the database as it checks if it should declare a user as 'active.' I 
still have to rewrite the purgeActive functions, and the command to set 
up the whole thing, but this is a good first step.