Rewrote purgeActive code to use database
Rewrote the purgeActive code to use the settings in the database. For 
efficiency's sake, I also removed the isActive function and integrated 
it into the purgeActiveMember function, which is something I previously 
did to the checkActive function 
(b01e47e769b9bf9dc8ca55b23ae6291c44f1543d). The end result is that now 
the entire active role feature is controlled from the database.