Expanded data cache to work on multiple servers
There are three objects in the code, used to store data that allows the 
active role management feature to operate. They're dict objects storing 
information indexed by the ids of various users. But after the most 
recent changes to the bot, namely the shifting of active role settings 
to a database, which would allow the system to work on multiple servers, 
I realized that merely indexing them by user id was not enough. I needed 
to store separate data depending on whether we're talking about a user's 
activity on one server or another. So it's now indexed by both server id 
and user id. Simple change, I think you'll find.