Wrote parseRole code
There are three different ways to specify a role in Discord. To deal 
with them, I wrote a new function that merely takes a server and a 
string and tries to extract a role from the server, defined by the 
string. I then decided to utilize it in the rolecall function, to 
broaden its use. I also altered the code so it no longer defines 'no 
role' as a 0, but as a 'None' object, as the new parseRole function 
does.