Reorganized a few constants
Was having a problem with the database file trying to be accessed from 
the wrong folder. So I had the brilliant idea to add those to the .env 
file. I also renamed the 'token' constant to match the style of all the 
other constant, and updated the instructions in the readme accordingly.