Migrated from SQLite to PostgreSQL
Made all the changes necessary to switch from using SQLite to 
PostgreSQL. Changed the database access library, and connection 
information. Updated all the prepared statements to use the standard 
Postgres scheme (%s) rather than a question mark. Updated the SQL file 
to be compatible with Postgres, updating many of the fields to use 
BIGINT data types, rather than INTEGER types, and changed the 
AUTOINCREMENT field to a SERIAL data type. And finally, updated the 
installation instructions to use Postgres instead of SQLite. And with 
all that, we're ready to use Postgres.