Added a few mod commands
Added three new mod commands, for kick, ban, and a special delete ban 
used to deal with spammers. This was written mainly because I'm sick of 
people using the standard ban command, which defaults to deleting 
messages. I've been burned way too many times by that. One time it was 
my ladyfriend making a joke. I had all three commands operate in similar 
ways. They record the reason for the kick/ban and send a message to the 
user explaining why they've been kicked/banned. As part of this, I added 
a parseMember function so I don't have to repeat the code for getting a 
member object from a string containing a mention. I only write it once 
and call the function. I also modified the help command to explain these 
new commands. But I learned that with this addition, I also went over 
the 2000 character limit. So I'm forced to also add the length control 
code I wrote for the rolecall function 
(d4b588c1f328457f787ac60f157734ef25204264). With that, everything is 
operational.