Undoing change to Procfile
Turns out I got it wrong in an earlier commit 
(12ee87bf735aec7ae62c8d2f5ef65c90c16958dd). Giving the startup command 
in the Procfile the 'release' label was causing multiple instances of 
the bot to run constantly. Each new deploy would add another instance, 
which is in many ways not good. So this should fix that. I did try 
setting the label to something descriptive, like 'MidnightBot', but that 
just caused nothing to happen at all. It seems the only label that's 
actually useful is 'worker'. It'll actually give me a toggle switch to 
turn the bot on and off and only run one instance. It's exactly what I 
need, but I can only wonder why it's the only one that'll work since 
there's no explanation in the Heroku documentation. Oh well...