Added the moderator name to kick/ban records
The audit log was not recording the names of whoever was responsible for 
a given kick/ban when using the command. So I fixed that. Doesn't change 
the output message, or the final message to the kicked/banned member. 
It's merely for the benefit of the ban record, and the audit logs.