Improved mute and unmute functions
Redesigned the mute and unmute functions to work more efficiently, while 
still remaining effective. I started by having a check to see if the 
member to be muted has full admin powers. In that case, muting would be 
useless, because full admin would bypass anything I could set in user 
permissions. So that prevents the bot from wasting its time. We then 
check to see if the member even has the ability to post or speak in the 
specified channels. If not, there's no reason to do anything. Similarly, 
the unmute function has similar checks to see if the user's permissions 
are even set to mute in a given channel. If not, there's no reason to do 
anything. If they were the only permissions set, we also have a check to 
clear out the whole overwrite, since it would no longer be necessary. 
All these checks allow the function to work more efficiently since it 
would be making fewer channel changes overall. Which means fewer records 
in the audit logs.