Redesigned database for payday settings
I had a bit of an epiphany while writing the code for the buyroles 
function. If one clears out the payday function, the role purchasing 
feature could still be set up and usable, but the way I designed the 
database, I couldn't get the name of the currency at that point. So I 
decided to create a separate table, just for the currency name. That way 
it'll remain, even if the payday function is removed.