Wrote setup commands for role purchasing
I began the first step in creating the role purchasing feature. I wrote 
the commands that allow administrators to set roles up for purchase, and 
also remove them from purchase. As a bonus, I also wrote a command for 
listing all roles on the purchase list. The command to actually buy 
roles though is still unfinished.