Replaced use of exceptions as conditional checks
To find if a user had the 'active role' in the 'checkActive' and 
'purgeActiveMember' functions, I was using exception handling, since if 
a user doesn't have a particular role, the 'index' method would throw a 
'ValueError' exception. But that's not really the proper way to handle 
things. So, I changed it, so it now checks using standard 'if' 
statements, and the 'count' method. It's much more reliable, fits best 
practices, and doesn't cause issues with error reporting if there are 
exceptions within the 'except' branches.