Shifted startup loop
The code that loops through the servers and runs basic startup 
procedures when the bot starts up has been taking way too long. Mainly 
it's the fault of the purgeActiveServer code. This results in the 
setupDataCache function not running before certain functions are called 
that require it. So, to sorta speed things up, that particular code has 
been split up into two separate loops. The functions that are urgent, 
and quick are run in the first loop, and the second loop handles 
functions that are slow and not so urgent, like the purgeActiveServer 
function.