Added activity record table
Added a special table in the database to record when a user was last 
considered 'active' by the standards of the 'checkActive' code. Added 
the necessary code to support the new table into the 'checkActive' 
function, and altered the function so it works on users who already have 
the role. I'm doing this because I want to change how the 'purgeActive' 
code works. But this is the first step in setting that up.