Changed names of all 'data' objects
The use of the name 'data' for objects containing database information, 
which is a very generic name, was getting frustrating. Since if I wanted 
to place another database fetch in the function, I would have to rename 
the original object, so it's clear there's a difference. So I changed 
all the names of the 'data' objects to be more descriptive, regardless 
of whether there's only one database call there or not.