Replaced use of 'now' method
The 'now' method from the 'datetime' class was all well and good for 
retrieving the current time. But it wasn't the best way, especially 
since we had a message object that had the current time already. So I 
decided to change it up. The time stored in the message object is more 
reliable and accurate since it's the time the actual command was sent, 
so better representative of the time the event being recorded happens. 
It's perfect.