Had purgeActive code look in the activity record
In an attempt to begin further streamlining the purgeActive code, I 
added a check into the activity record table I created earlier 
(0aeecac736964bcd3c1e47f6b9497469803a6d42). If the code finds anything 
in that table, it uses that over any history checks. This means members 
will need to fulfill the 'active' criteria within the 'max inactivity' 
period, to prevent the role being taken away. Merely making a single 
post is not enough anymore. The original criteria will still work if 
there is no activity record in the database, but if and when I remove 
the history checks, that'll no longer be the case.