Fixed datatype issue with activity record
I forgot to convert the stored date timestamp in the activity record 
table, into a datetime object, when pulling the data out of the 
database. I noticed this once it threw an exception because I tried 
comparing a datetime object with an int. The fix was quite simple, just 
using the fromtimestamp method. And with that, everything involving that 
database fetch should work now.