Continued to streamline the purgeActive code
As part of my continued efforts to streamline the purgeActive code, I 
removed the second loop and placed its code in the first loop. This is 
mainly because I could, and because it streamlines the process and 
removes the need for a separate list object. I also decided having a 
check to see if 'lastMessageTime is None' is really a bit unnecessary, 
especially since I could just, instead, initialize 'lastMessageTime' as 
a datetime object set to the minimum value, since I'm trying to find the 
highest datetime value in the loop initializing it to its minimum value 
will not interfere with that. This also allows me to remove the final 
'is None' check, because the initialization guarantees it'll never equal 
None.