Removed tbl_accounts
The table, tbl_accounts, was really just a summary table. It contained 
nothing more, nor less than a summary of what was also in 
tbl_transactions. So it was a bit redundant. Which is why I decided to 
remove it and replace all calls to it with comparable calls to 
tbl_transactions. This works much better.