Wrote more persistence code for role buying
Wrote a new function for role persistence with the buyable roles, this 
time focusing on one role at a time, rather than one member. So far, 
only used when a new role is set up to be buyable. If it's never been 
buyable before, it'll clear out everyone who has the role. However, If 
there is any sign in the transaction record of this role being bought 
before, those that did will either be granted the role or allowed to 
keep it. Should be pretty effective.