Fixed error in payday function
The payday function was crashing on the first payday because of a small 
oversight. After I fetched the necessary data from the transaction 
table, for each statement I checked if the whole row was a None object, 
rather than the first element, which is what I should've done since 
that's how the MAX and SUM statements work. They return proper row 
objects regardless, but if there's no data, the individual fields will 
be None, not the whole row.