Added an error message for server members
Whenever an exception is raised, it sends me a DM alerting me to the 
issue. However, if it's caused by a command, the user who made the 
command gets no feedback. So I added a bit of code to the on_error 
function to output some kind of message to whatever channel it could 
trace the error to. True, it's not like the users could do anything 
about it, but at least they'd know something happened... some kind of 
error. Even if they don't know what the error is. It also means they 
won't start spamming the command assuming they're just not being 
persistent enough.