Removed old code from purgeActive
The code for selecting who no longer qualifies for the 'active' role was 
based on the last message posted. A while ago, I decided to retool it so 
its based on when they last qualified as 'active' 
(64e9a178e43b124cfe8ab826b8c8589baf40d708). But I kept the old code 
around as a fallback for legacy reasons. Eventually, I planned to simply 
disable the code with a switch, but keep the code around, in case I 
needed to reenable it (c787749dd22501fbd3394f49e423fa2431e8d53a). But 
now, I'm just going to remove it altogether. It's still in the git 
record, and I can, if I really want, shove it back in at a later date. 
But simply removing the old code, and restructuring the entire code 
block for efficiency is a much better solution.