Added a few conditions to the rules management feature
On any server where the rule management feature wasn't implemented, the 
two commands for getting a copy of the rules sent to your DMs would send 
an empty rule list; just a title, and nothing else. This is pointless. 
So I disabled the commands if they can't find any rules to actually 
send. I also added some conditions to the help feature. So the commands 
for the rule management feature won't be listed to users if there are no 
rules recorded, with the exception being the setup commands for admins. 
They'll still be listed even if there are no rules in the system.