Reorganized mute and unmute functions
Reorganized the mute and unmute functions to mute or unmute users in the 
categories first, before continuing in the subchannels. This is because, 
rather than provide the channels in a logical order (like the order of 
their position in the channel list) the guild.channels property gives 
them in a seemingly random order (likely the order in which they were 
created) which isn't really helpful. And by setting the mutes/unmutes in 
the categories first, the functions won't set them in any synced 
channels, reducing the amount of work it has to do, and the number of 
reports it'll place in the audit log.