Expanded error reporting
In an expansion to the error reporting for server members 
(5640f2a7f61e530c82db6ec712fa5f2f85593145) I added a hook that allows 
Midnight to report error messages, even if it doesn't have posting 
permission in the relevant channel. In this case, it reports it to the 
user's DMs. I also added a hook so I know when Midnight's been blocked 
by a given user, just for my own sake.