Altered how the refundRole function works
Previously, the refundRole function worked by searching for the role in 
the price list and refunding the cost from there. But generally, that's 
not how refunds work. If the price changes, it doesn't matter, the money 
you get back is the money you paid. So I altered the code so that it 
searches for the price one paid in the transaction record and refunds 
that. It's more accurate and more fair.