Added some remote admin commands
Added a couple of commands that allow me to, from DMs, list which 
servers Midnight is running on, and pull her from select servers if I 
think it's necessary. I also programmed it to report to me whenever 
Midnight joins a new server. This may seem like an overreach, but it 
allows me to stop anyone from abusing her, or her resources.