Modified channel list output
The channel list wasn't being properly displayed. Instead of being in a 
logical order, it was in a seemingly random order. I eventually figured 
out that the 'position' property didn't mean what I thought it meant, 
and it wouldn't sort them in the order I thought it would. So I decided 
to use a completely different function, the by_category function, which 
is a much better way of sorting through the channels. I had to alter the 
rest of the code to fit, but it works much better now.