Expanded mute and unmute to specify channel
The mute and unmute commands can now specify specific channels for the 
user to be muted and unmuted on. So if one's only being a nuisance on 
one channel, this command will allow mods to restrict them on that 
specific channel. I also added an explanation of this new feature to the 
output of the help command.