Added new commands for links
Two new commands for two different links. One is to get Midnight's 
source code, which I removed from the help output. The other is to get 
an invite to Midnight's Sanctuary, a new server I created for the bot. I 
just started it, but I intend for it to be the place to discuss bugs, 
new features, and troubleshooting tips.