Wrote Ghost Ping Detector
Added a new feature to detect when someone does a 'ghost ping', which is 
a ping of @everyone or @here that is subsequently deleted, and posts a 
report of said 'ghost ping' to the channel in which it occured. Also 
added a feature that detects when the reports are deleted, and reposts 
them, with added notes. This is mainly to deal with issues on Ponyworld, 
and is the second time I had to do something like this. See also: YAG 
Snipe (53b604fd4ae40b22031eb5e8f811ef37ccda70f6).