Modified Ghost Ping Detector
Added a new case for the Ghost Ping Detector to detect. Because many 
servers have one role that all members have, some people might try to 
ghost ping those roles instead. The code now considers any role that at 
least 75% of members have to fall into the detector's parameters, and if 
someone decides to ghost ping those roles, the detector will issue a 
report as if it was an @everyone ping.