Accounted for add_reactions permission in mute
The mute command was allowing people to continue to add emoji reactions 
to messages, which can get annoying, and undermines a major point of the 
mute, that they're supposed to be completely silenced. So I added the 
'add_reactions' permission to the list of permissions to strip. It 
should be quite effective.