Fixed active role race conditions
Similar to the issue I encountered with buyable roles 
(55d3685f53281e63ded0e53524b3bd1defa12820), the active role code was 
having problems with race conditions, where the assigning of the role 
was triggering checks before the database was updated, and the old 
information in the database would cause the role to be taken away. By 
moving the assigning of the role below the point where the database 
update happens, I fix this problem.