Fixed issue with deleted roles
If a paid role is deleted without being removed from the database first, 
it causes errors, mainly in the 'rolemenu' command. So I set it up to 
delete those roles from the database the moment that happens, or the 
next time the 'rolemenu' command is run (as a fallback). I also set the 
persist function to just ignore the deleted roles. Not because any 
problems were being caused, but because in theory, problems could be 
caused if more than one buyable role was deleted. I'm not sure though, 
this is just a precaution.