Merged purge and persistActive functions
Because they basically serve the same function, just in opposite 
directions, I decided to merge the two functions for purging and 
persisting the active role. A lot of the code is the same, so it just 
made sense. I also removed the purgeActiveServer function, since there's 
no need to run a full scan of every server every time the bot starts, 
and it could get cumbersome as the bot's adoption grows.