Active role persist affects new members
There was a bit of a flaw in the code that enforces the active role 
persistence. It's supposed to not only grant the role if the user 
qualifies for it based on the activity record but take it away if they 
don't. And the code was written so a new member, with nothing in the 
activity record, would not be affected by this code. This means they 
would not have it taken away if they had it granted manually, which 
should happen since they would fail to qualify. So I made a quick 
change, where if the member has nothing in the activity record, the code 
will assume they last qualified for the role at the beginning of time, 
and therefore take the role away if it's been granted. This solves that 
problem.