Added Unicode emoji detection to emoji censor
The emoji censor now detects Unicode emojis. There are a few problems 
with it I haven't been able to nut out, but so far this seems effective. 
It utilizes a bit of regex I was able to put together from some advice 
on the discord.py Discord server, and my own investigations by looking 
through some Unicode tables. This should be enough to deal with your 
standard emoji spam across the board.