Created a new promoter role feature
The new promoter role feature assigns a role to any member who has 
recruited enough new members. The bot only counts members it sees 
arriving, so if it's offline when new members arrive, it won't count 
those members. Nonetheless, this is an effective means of rewarding the 
promotion of the server.