Prevent storing of invites when promoter role is not set up
The promoter role code was storing invites even if there was no need to 
because the promoter role was not set up. So, I made a quick change. It 
checks if the role is set up before storing anything in the cache. It 
also runs the data cache setup after the role is set up, and clears it 
out if the role is cleared out. This is especially necessary since the 
cache is only updated when new members join, and won't be updated at all 
during the time the role setup has been cleared out.