Changed Postgre driver
Switched from using psycopg2 to asyncpg. It should be more efficient, 
thanks to its use of multithreading. This was suggested by 'ehrenbruder' 
on Discord, and I looked into it, and it seems like the best idea. I 
went through a lot of trouble to get this working since it uses a 
significantly different syntax, but it's clearer and more concise. I 
prefer it.