Fixed error in raised exception
The Exception that's raised in the timeDeltaToString had a bit of an 
issue. I forgot to convert the length of the totalTimeParts array into a 
string, so it can be outputted. Ended up causing a problem when that 
very Exception ended up getting raised and I couldn't tell why. So 
that's been fixed.