Fixed issue with recruiting records
There was an issue with the recordRecruit code. The code was written so 
it can detect which invite the new member used by comparing the invites 
in the cache with the invites that are currently in play and comparing 
the number of uses in each. Whichever one has been used more since the 
last check of the cache must be the one that was used. As part of this, 
I wrote that in the event the appropriate invite could not be found, 
that we scan through the current slate of invites, because they were 
probably created since. Unfortunately, this caused problems, because any 
invite that only had one use would not be found by this code, because it 
would cease to exist the moment it was used. I found the solution was to 
go through the audit log and update the cache with any invites that had 
been created since the cache was last updated, as well as delete any 
that were deleted. That way I would catch any that only had one use. 
Then going through the list, same as before, with additional checks for 
those that reached their max uses, or have expired. I also altered how 
the invite cache works. Rather than store the actual invite object, it 
just stores the necessary information as a tuple, in a dictionary, with 
the invite code as the key. Most effective way of doing things, 
especially now that we're working through the audit log since those 
invite objects won't be very reliable in that case.