Altered how invite cache is updated
I had a few exceptions which I think were caused by multiple users 
joining a server using the same invite in rapid succession. This, in 
theory, would result in the cache being updated with the invite's 
current status, after it's been used several times, while the first use 
is still being processed. So it would be impossible to trace the origin 
invite of the members that follow; which results in the exception. I 
fixed this by making sure the number of uses recorded in the cache 
increment one at a time, and not ever with three at a time. It should 
fix the problem. If it keeps happening... I don't know.