Will now delete invites from banned members
Any member who's banned was probably banned for a good reason. To 
prevent these people from inviting their alts, or friends to raid the 
server, I decided to include code that immediately deletes all invites 
that had been created by the banned member. It could result in issues if 
said member also promoted the server on their own server, as a bit of 
cross-promotion, but there's not much I can do to stop that.