Wrote timeout command
A new timeout command, once set up, allows moderators to segregate 
misbehaving members to a single specific channel. This allows them to be 
talked with privately by staff, until such time that the user either 
calms down, apologizes, or is straight-up kicked or banned. It also 
includes persistence, so the timeout can't be circumvented by leaving 
and rejoining.