Altered datetime datatypes in database
Using integer datatypes to store datetime and timedelta objects doesn't 
make a lot of sense when there are more specialized datatypes I could 
use. So I changed the appropriate columns in the database to use 
TIMESTAMP or INTERVAL datatypes. One major benefit of this change is 
that I no longer have to run conversion to and from integer and datetime 
or timedelta objects, so I removed all the now-obsolete data conversion 
code. But the one hiccup in this scheme is that it's not as simple as 
just replacing the database. Any data already stored in the system now 
has to be converted to the new datatypes. This is actually really 
simple, and only requires a bit of SQL code. But I decided to write some 
python code to do it, not only to run the migration, but to test if the 
migration was successful, complete, and accurate. I placed the migration 
script in a special folder. Any future migration scripts will be stored 
there as well.