Simplified mute and timeout features with roles
Added the ability to designate mute and timeout roles, which will be 
used to simplify and speed up the mutes and timeouts. The 
channel-by-channel code will still run as designed. But it won't have to 
run on quite as many channels since the role should take care of that 
for us. The channel-by-channel code is just to fill in any accidental 
gaps.