Fixed timeout persist
The timeout persist was never properly tested. I just tested it and 
found out there were multiple problems. The first being the fact that 
the timeouts were not being stored in the database because the 
conditional statement was incorrectly written, and the INSERT statement 
referenced the wrong table. The next issue was simpler; a few incorrect 
variables from copy-pasted code. And finally, there's the fact that the 
persistTimeout function wasn't even being called in the main python 
script. That's all been corrected.