Fixed timeout role setup
There was a slight oversight when designing the code to build the 
timeout role (e78e51b25b87a1fd2d379efe11627ec8f8150e50). In order to 
simplify the design of the code for setting the role permissions in 
channels, I used the same function used for setting member permissions 
in channels, but added a hook that primes the code by saying all 
permissions are set, when it's a role, not a member, so the code 
properly denies the permissions, as it's supposed to. However, the 
channelOpenForTimeout function is supposed to allow permissions, not 
deny them. As a result, the code assumes the permissions are already 
allowed, and doesn't run. So I changed it so the function is primed by 
saying no permissions are set, which means the code actually runs now, 
as I want it to.