Dealt with double-space glitch
There was a slight glitch when one tries to mute or timeout any user and 
accidentally puts extra spaces in the command, which is a common mistake 
if you're on mobile, due to how the mobile interface works. I threw in a 
few extra bits of code to deal with it, mainly by stripping out the 
extra whitespaces when they're likely to cause problems. So far, I'm 
only doing it for the moderation commands, since they're the most 
important, and the utility functions because they're used by the 
moderation commands. I plan to implement a much more comprehensive fix, 
but it involves a complete redesign in how commands are handled. So, 
this will do for now.