Fixed timing issue with ActiveRole
There was a slight problem with the active role feature. The max time 
threshold was being calculated using local time, which, on the host I'm 
running Midnight on, is five hours behind UTC. But it's being compared 
to the recorded output of the created_at property, which is always in 
UTC. As a result, the comparison is five hours off. So I changed it so 
the whole thing uses UTC, for consistency. This really only has a major 
impact if the max time is only a few hours or a few minutes, which 
shouldn't be that common. But for the sake of consistency and accuracy, 
I still think it's a good idea.