Changed the variable names in channel mute and timeout functions
Since the functions of the channelMute, channelBanishForTimeout, and 
channelOpenForTimeout functions were expanded for the role-based 
enhancement (e78e51b25b87a1fd2d379efe11627ec8f8150e50), using the term 
'member' to refer to the action's primary subject didn't make a lot of 
sense. So I changed it to 'subject.' It's a minor change, and it's more 
for style purposes, but I think it's necessary.