Used try-finally on database calls
We were experiencing an issue where the entire bot would lock up after 
any event where a series of exceptions were triggered. I eventually 
sussed that it was being caused by the connection pool, and connections 
not being returned due to exceptions being thrown. The best solution to 
this is to place all sections with database calls in try-finally blocks, 
so no matter what, the connection would eventually be returned to the 
pool. It's what I should've done from the very beginning.