Updated permissions on invite
Administrator permissions are required for the bot to properly set 
permissions in all the channels for the mute and timeout features... 
that is unless Midnight is granted the required permissions manually. 
Nonetheless, I do think Midnight should be granted admin permissions for 
the sake of simplicity. So I added administrator permissions to the 
invite link. Despite the fact that all the other permissions are now 
rendered redundant by the admin permissions, I set the link to still set 
all the other permissions in case the server admins don't want to give 
full admin access or decide to randomly disable it at a later date. I 
also included a note explaining this, that disabling full admin access 
is an option.