Updated feature list
Updated the feature list in the readme. As part of this, I added a 
separate section for features that are coded, and were once active, but 
have since been disabled; such as the two snipe functions, the ghost 
ping detector, and the emoji spam detector. I also added the timeout 
command and the promoter role to the list of working features. Finally, 
I also added what I'm calling a 'bumper role' to the upcoming features 
list.