Added new bumper role feature
New 'bumper role' feature keeps track of who bumps the server on 
Disboard, based on whether they successfully ran the bump command on the 
server, and grants special roles based on how many times it worked for 
them. This whole thing was relatively easy to implement, the only thing 
that was new to me was, I had to examine how the embeds worked, so I 
could properly analyze them for the important confirmation message. I 
also included a scan command to scan a particular channel for any server 
bumps, to allow the system to count bumps that happened before the role 
was initially set up. I wasn't able to fully test this, due to the 
limitation of my setup. The granting of roles immediately after a 
successful bump couldn't be tested properly because of the bump 
cooldown. But if it doesn't work, the persist should fill in the gaps, 
and I'll just correct the error later.