Redesigned economy data storage
The database was getting unnecessarily bloated, what with me storing 
every single transaction in it when most of the information was being 
stored and not directly accessed. So I decided to make it more efficient 
by changing it from being transaction-based, to merely storing a summary 
of the data. Just a record of what's been purchased, and what the 
member's current balance is. The transactions are still stored, but not 
in the database, instead, they are now stored in log files, in a CSV 
format. This involved rewriting a lot of code to remove any use of the 
transactions table and replace them with the use of the two new tables. 
I also wrote some migration code to handle the changeover, and this time 
it also accesses discord to fetch some information for the headers in 
the transaction log files.